+++ core/src/main/scala/astraea/spark/rasterframes/extensions/RFSpatialColumnMethods.scala
@@ -103,7 +103,6 @@ trait RFSpatialColumnMethods extends MethodExtensions[RasterFrame] with Standard
       case rf ⇒ rf.certify
     }
   }
-
 }
 
 object RFSpatialColumnMethods {