+++ core/src/main/scala/astraea/spark/rasterframes/extensions/RasterFrameMethods.scala
@@ -16,13 +16,13 @@
 
 package astraea.spark.rasterframes.extensions
 
-import java.sql.Timestamp
 import java.time.ZonedDateTime
 
+import astraea.spark.rasterframes.util._
 import astraea.spark.rasterframes.{MetadataKeys, RasterFrame}
 import geotrellis.proj4.CRS
 import geotrellis.raster.resample.{Bilinear, ResampleMethod}
-import geotrellis.raster.{MultibandTile, ProjectedRaster, Tile, TileLayout}
+import geotrellis.raster.{CellGrid, MultibandTile, ProjectedRaster, Tile, TileLayout}
 import geotrellis.spark._
 import geotrellis.spark.io._
 import geotrellis.spark.tiling.{LayoutDefinition, Tiler}
@@ -32,11 +32,8 @@ import org.apache.spark.annotation.Experimental
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.gt.types.TileUDT
-import org.apache.spark.sql.types.{Metadata, StructField, TimestampType}
+import org.apache.spark.sql.types.{Metadata, TimestampType}
 import spray.json._
-import astraea.spark.rasterframes.util._
-import astraea.spark.rasterframes.encoders.StandardEncoders._
 
 import scala.reflect.runtime.universe._
 
@@ -46,57 +43,36 @@ import scala.reflect.runtime.universe._
  */
 trait RasterFrameMethods extends MethodExtensions[RasterFrame]
   with RFSpatialColumnMethods with MetadataKeys with LazyLogging {
-  import Implicits.WithDataFrameMethods
-  import Implicits.WithRasterFrameMethods
+  import Implicits.{WithDataFrameMethods, WithRasterFrameMethods}
 
   private val _stableDF = self
   import _stableDF.sqlContext.implicits._
 
-  /** Get the names of the columns that are of type `Tile` */
-  def tileColumns: Seq[TypedColumn[Any, Tile]] =
-    self.schema.fields
-      .filter(_.dataType.typeName.equalsIgnoreCase(TileUDT.typeName))
-      .map(f ⇒ col(f.name).as[Tile])
+
+  /**
+   * A convenience over `DataFrame.withColumnRenamed` whereby the `RasterFrame` type is maintained.
+   */
+  def withRFColumnRenamed(existingName: String, newName: String): RasterFrame =
+    (self: DataFrame).withColumnRenamed(existingName, newName).certify
 
   /** Get the spatial column. */
   def spatialKeyColumn: TypedColumn[Any, SpatialKey] = {
-    val key = findSpatialKeyField
+    val key = self.findSpatialKeyField
     key
       .map(_.name)
       .map(col(_).as[SpatialKey])
       .getOrElse(throw new IllegalArgumentException("All RasterFrames must have a column tagged with context"))
   }
 
-  /** Get the temporal column, if any. */
-  def temporalKeyColumn: Option[TypedColumn[Any, TemporalKey]] = {
-    val key = findTemporalKeyField
-    key.map(_.name).map(col(_).as[TemporalKey])
-  }
-
-  private[rasterframes] def findRoleField(role: String): Option[StructField] =
-    self.schema.fields.find(
-      f ⇒
-        f.metadata.contains(SPATIAL_ROLE_KEY) &&
-          f.metadata.getString(SPATIAL_ROLE_KEY) == role
-    )
-
-  /** The spatial key is the first one found with context metadata attached to it. */
-  private[rasterframes] def findSpatialKeyField: Option[StructField] =
-    findRoleField(SPATIAL_KEY_COLUMN.columnName)
-
-  /** The temporal key is the first one found with the temporal tag. */
-  private[rasterframes] def findTemporalKeyField: Option[StructField] =
-    findRoleField(TEMPORAL_KEY_COLUMN.columnName)
-
   /**
    * Reassemble the [[TileLayerMetadata]] record from DataFrame metadata.
    */
   def tileLayerMetadata: Either[TileLayerMetadata[SpatialKey], TileLayerMetadata[SpaceTimeKey]] = {
-    val spatialMD = findSpatialKeyField
+    val spatialMD = self.findSpatialKeyField
       .map(_.metadata)
       .getOrElse(throw new IllegalArgumentException(s"RasterFrame operation requsted on non-RasterFrame: $self"))
 
-    if (findTemporalKeyField.nonEmpty)
+    if (self.findTemporalKeyField.nonEmpty)
       Right(extract[TileLayerMetadata[SpaceTimeKey]](CONTEXT_METADATA_KEY)(spatialMD))
     else
       Left(extract[TileLayerMetadata[SpatialKey]](CONTEXT_METADATA_KEY)(spatialMD))
@@ -107,7 +83,7 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
 
   /** Add a temporal key to the RasterFrame, assigning the same temporal key to all rows. */
   def addTemporalComponent(value: TemporalKey): RasterFrame = {
-    require(temporalKeyColumn.isEmpty, "RasterFrame already has a temporal component")
+    require(self.temporalKeyColumn.isEmpty, "RasterFrame already has a temporal component")
     val tlm = tileLayerMetadata.left.get
     val newTlm = tlm.map(k ⇒ SpaceTimeKey(k, value))
 
@@ -204,8 +180,8 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
    */
   def clipLayerExtent: RasterFrame = {
     val metadata = tileLayerMetadata
-    val extent = metadata.fold(_.extent, _.extent)
-    val layout = metadata.fold(_.layout, _.layout)
+    val extent = metadata.widen.extent
+    val layout = metadata.widen.layout
     val trans = layout.mapTransform
 
     def updateBounds[T: SpatialComponent: Boundable: JsonFormat: TypeTag](tlm: TileLayerMetadata[T],
@@ -216,16 +192,16 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
 
       val gridExtent = trans(keyBounds.toGridBounds())
       val newExtent = gridExtent.intersection(extent).getOrElse(gridExtent)
-      self.setSpatialColumnRole(spatialKeyColumn, tlm.copy(extent = newExtent, bounds = keyBounds))
+      self.setSpatialColumnRole(self.spatialKeyColumn, tlm.copy(extent = newExtent, bounds = keyBounds))
     }
 
     val df = metadata.fold(
-      tlm ⇒ updateBounds(tlm, self.select(spatialKeyColumn)),
+      tlm ⇒ updateBounds(tlm, self.select(self.spatialKeyColumn)),
       tlm ⇒ {
         updateBounds(
           tlm,
           self
-            .select(spatialKeyColumn, temporalKeyColumn.get)
+            .select(self.spatialKeyColumn, self.temporalKeyColumn.get)
             .map { case (s, t) ⇒ SpaceTimeKey(s, t) }
         )
       }
@@ -240,10 +216,10 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
    */
   def toTileLayerRDD(tileCol: Column): Either[TileLayerRDD[SpatialKey], TileLayerRDD[SpaceTimeKey]] =
     tileLayerMetadata.fold(
-      tlm ⇒ Left(ContextRDD(self.select(spatialKeyColumn, tileCol.as[Tile]).rdd, tlm)),
+      tlm ⇒ Left(ContextRDD(self.select(self.spatialKeyColumn, tileCol.as[Tile]).rdd, tlm)),
       tlm ⇒ {
         val rdd = self
-          .select(spatialKeyColumn, temporalKeyColumn.get, tileCol.as[Tile])
+          .select(self.spatialKeyColumn, self.temporalKeyColumn.get, tileCol.as[Tile])
           .rdd
           .map { case (sk, tk, v) ⇒ (SpaceTimeKey(sk, tk), v) }
         Right(ContextRDD(rdd, tlm))
@@ -252,10 +228,14 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
 
   /** Convert all the tile columns in a Rasterrame to a GeoTrellis [[MultibandTileLayerRDD]] */
   def toMultibandTileLayerRDD: Either[MultibandTileLayerRDD[SpatialKey], MultibandTileLayerRDD[SpaceTimeKey]] =
+    toMultibandTileLayerRDD(self.tileColumns: _*)
+
+  /** Convert the specified tile columns in a Rasterrame to a GeoTrellis [[MultibandTileLayerRDD]] */
+  def toMultibandTileLayerRDD(tileCols: Column*): Either[MultibandTileLayerRDD[SpatialKey], MultibandTileLayerRDD[SpaceTimeKey]] =
     tileLayerMetadata.fold(
       tlm ⇒ {
         val rdd = self
-          .select(spatialKeyColumn, array(self.tileColumns: _*)).as[(SpatialKey, Array[Tile])]
+          .select(self.spatialKeyColumn, array(tileCols: _*)).as[(SpatialKey, Array[Tile])]
           .rdd
           .map { case (sk, tiles) ⇒
             (sk, MultibandTile(tiles))
@@ -264,7 +244,7 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
       },
       tlm ⇒ {
         val rdd = self
-          .select(spatialKeyColumn, temporalKeyColumn.get, array(self.tileColumns: _*)).as[(SpatialKey, TemporalKey, Array[Tile])]
+          .select(self.spatialKeyColumn, self.temporalKeyColumn.get, array(tileCols: _*)).as[(SpatialKey, TemporalKey, Array[Tile])]
           .rdd
           .map { case (sk, tk, tiles) ⇒ (SpaceTimeKey(sk, tk), MultibandTile(tiles)) }
         Right(ContextRDD(rdd, tlm))
@@ -275,8 +255,50 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
   private[rasterframes] def extract[M: JsonFormat](metadataKey: String)(md: Metadata) =
     md.getMetadata(metadataKey).json.parseJson.convertTo[M]
 
-  /** Convert the tiles in the RasterFrame into a single raster. For RasterFrames keyed with temporal keys, this
-    * will merge based */
+  // TODO: Take care of DRY below
+//  private def rasterize[T <: CellGrid: TypeTag](
+//    tileCols: Seq[Column],
+//    rasterCols: Int,
+//    rasterRows: Int,
+//    resampler: ResampleMethod): ProjectedRaster[T] = {
+//
+//    val clipped = clipLayerExtent
+//
+//    val md = clipped.tileLayerMetadata.widen
+//    val newLayout = LayoutDefinition(md.extent, TileLayout(1, 1, rasterCols, rasterRows))
+//
+//    val trans = md.mapTransform
+//
+//    //val cellType = rdd.first()._2.cellType
+//    val keyBounds = Bounds(SpatialKey(0, 0), SpatialKey(0, 0))
+//    val newLayerMetadata =
+//      md.copy(layout = newLayout, bounds = keyBounds)
+//
+//
+//    val newLayer = typeOf[T] match {
+//      case tpe if tpe <:< typeOf[Tile] ⇒
+//        val r = clipped.toTileLayerRDD(tileCols.head)
+//          .fold(identity, _.map { case (stk, t) ⇒ (stk.spatialKey, t) }) // <-- Drops the temporal key outright
+//          .map { case (key, tile) ⇒ (ProjectedExtent(trans(key), md.crs), tile) }
+//        ContextRDD(r, md)
+//          .tileToLayout(newLayerMetadata, Tiler.Options(resampler))
+//      case tpe if tpe <:< typeOf[MultibandTile] ⇒
+//        val r = clipped.toMultibandTileLayerRDD(tileCols: _*)
+//          .fold(identity, _.map { case (stk, t) ⇒ (stk.spatialKey, t) }) // <-- Drops the temporal key outright
+//          .map { case (key, tile) ⇒ (ProjectedExtent(trans(key), md.crs), tile) }
+//        ContextRDD(r, md)
+//          .tileToLayout(newLayerMetadata, Tiler.Options(resampler))
+//    }
+//
+//    val stitchedTile = newLayer.stitch()
+//
+//    val croppedTile = stitchedTile.crop(rasterCols, rasterRows)
+//
+//    ProjectedRaster(croppedTile, md.extent, md.crs)
+//  }
+
+  /** Convert the tiles in the RasterFrame into a single raster. For RasterFrames keyed with temporal keys, they
+    * will be merge undeterministically. */
   def toRaster(tileCol: Column,
                rasterCols: Int,
                rasterRows: Int,
@@ -286,10 +308,10 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
 
     val md = clipped.tileLayerMetadata.widen
     val trans = md.mapTransform
-    val keyCol = clipped.spatialKeyColumn
     val newLayout = LayoutDefinition(md.extent, TileLayout(1, 1, rasterCols, rasterRows))
 
-    val rdd: RDD[(SpatialKey, Tile)] = clipped.select(keyCol, tileCol).as[(SpatialKey, Tile)].rdd
+    val rdd = clipped.toTileLayerRDD(tileCol)
+      .fold(identity, _.map{ case(stk, t) ⇒ (stk.spatialKey, t) }) // <-- Drops the temporal key outright
 
     val cellType = rdd.first()._2.cellType
 
@@ -310,4 +332,39 @@ trait RasterFrameMethods extends MethodExtensions[RasterFrame]
     ProjectedRaster(croppedTile, md.extent, md.crs)
   }
 
+  /** Convert the Red, Green & Blue assigned tiles in the RasterFrame into a single color composite raster.
+   * For RasterFrames keyed with temporal keys, they will be merged underterministically. */
+  def toMultibandRaster(
+    tileCols: Seq[Column],
+    rasterCols: Int,
+    rasterRows: Int,
+    resampler: ResampleMethod = Bilinear): ProjectedRaster[MultibandTile] = {
+
+    val clipped = clipLayerExtent
+
+    val md = clipped.tileLayerMetadata.widen
+    val trans = md.mapTransform
+    val newLayout = LayoutDefinition(md.extent, TileLayout(1, 1, rasterCols, rasterRows))
+
+    val rdd = clipped.toMultibandTileLayerRDD(tileCols: _*)
+      .fold(identity, _.map{ case(stk, t) ⇒ (stk.spatialKey, t)}) // <-- Drops the temporal key outright
+
+    val cellType = rdd.first()._2.cellType
+
+    val newLayerMetadata =
+      md.copy(layout = newLayout, bounds = Bounds(SpatialKey(0, 0), SpatialKey(0, 0)), cellType = cellType)
+
+    val newLayer = rdd
+      .map {
+        case (key, tile) ⇒
+          (ProjectedExtent(trans(key), md.crs), tile)
+      }
+      .tileToLayout(newLayerMetadata, Tiler.Options(resampler))
+
+    val stitchedTile = newLayer.stitch()
+
+    val croppedTile = stitchedTile.crop(rasterCols, rasterRows)
+
+    ProjectedRaster(croppedTile, md.extent, md.crs)
+  }
 }